backend/requirements.txt
tensorflow
numpy
pillow
flask
flask-cors
python-dotenv
openai
requests
